Rename bitrate to datarate, bandwidth to frequency_deviation for FSK.
The bitrate to datarate is backwards compatible as it does not change
the Protobuf field number. The bandwidth was never used as this value is
not available in the LoRa HAL on FSK uplink / downlink. To send a FSK
downlink, both the datarate and freq. deviation must be provided (which
is not happening currently).