Implement sending multiple downlink possibilities to gw.
This allows for sending the RX1 and RX2 downlink option at once to the
gateway. Note that each downlink option has its own phy_payload field as
due to the different data-rates, the NS might include a different set of
mac-commands in the phy_payload to comply with max. payload
restrictions.